move confirmform to separate folder + revert signup.tsx
