+++ src/components/Signup/ConfirmForm/ConfirmForm.tsx
@@ -3,12 +3,13 @@ import { AuthError, confirmSignUp, resendSignUpCode } from "aws-amplify/auth";
 import { Form } from "components/form";
 import { GENERIC_ERROR_MESSAGE } from "constants/common";
 import { useErrorContext } from "contexts/ErrorContext";
-import { useEffect, useRef, useState } from "react";
+import { useState } from "react";
 import { UseFormReturn, useForm } from "react-hook-form";
 import { requiredString } from "schemas/string";
 import { object } from "yup";
-import { CodeRecipientEmail, StateSetter, UserType } from "./types";
-import Field from "./Field";
+import Field from "../Field";
+import { CodeRecipientEmail, StateSetter, UserType } from "../types";
+import ResendOTPTimer from "./ResendOTPTimer";
 
 type Props = {
   codeRecipientEmail: CodeRecipientEmail;
@@ -30,49 +31,49 @@ export default function ConfirmForm(props: Props) {
 
   type FV = typeof methods extends UseFormReturn<infer U> ? U : never;
 
-  const [remainingTime, setRemainingTime] = useState(2);
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setRemainingTime((prev) => {
-        if (prev === 1) {
-          clearInterval(interval);
-        }
-        return prev - 1;
+  const submit = async (fv: FV) => {
+    try {
+      const result = await confirmSignUp({
+        username: props.codeRecipientEmail.raw,
+        confirmationCode: fv.code,
       });
-    }, 1000);
 
-    return () => {
-      clearInterval(interval);
-    };
-  }, []);
+      if (result.nextStep.signUpStep !== "DONE" || !result.isSignUpComplete) {
+        return handleError("Code confirmation failed");
+      }
+
+      props.setSignupState({ type: "success", userType: props.userType });
+    } catch (err) {
+      const message =
+        err instanceof AuthError ? err.message : GENERIC_ERROR_MESSAGE;
+      handleError(message);
+    }
+  };
+
+  const resendOTP = async () => {
+    try {
+      setIsRequestingNewCode(true);
+      //no need to inspect result
+      await resendSignUpCode({
+        username: props.codeRecipientEmail.raw,
+      });
+
+      return alert("New code has been sent to your email.");
+    } catch (err) {
+      const message =
+        err instanceof AuthError ? err.message : GENERIC_ERROR_MESSAGE;
+      handleError(message);
+    } finally {
+      setIsRequestingNewCode(false);
+    }
+  };
 
   return (
     <Form
       className={`${props.classes ?? ""} grid text-[#1D3C51]`}
       disabled={isSubmitting || isRequestingNewCode}
       methods={methods}
-      onSubmit={handleSubmit(async (fv) => {
-        try {
-          const result = await confirmSignUp({
-            username: props.codeRecipientEmail.raw,
-            confirmationCode: fv.code,
-          });
-
-          if (
-            result.nextStep.signUpStep !== "DONE" ||
-            !result.isSignUpComplete
-          ) {
-            return handleError("Code confirmation failed");
-          }
-
-          props.setSignupState({ type: "success", userType: props.userType });
-        } catch (err) {
-          const message =
-            err instanceof AuthError ? err.message : GENERIC_ERROR_MESSAGE;
-          handleError(message);
-        }
-      })}
+      onSubmit={handleSubmit(submit)}
     >
       <h3 className="text-center text-2xl font-bold text-navy-d4">
         Verify your account
@@ -95,29 +96,7 @@ export default function ConfirmForm(props: Props) {
 
       <span className="flex-center gap-1 text-sm font-medium mt-5">
         Resend code in
-        <button
-          type="button"
-          className="text-blue-d1 hover:text-blue active:text-blue-d2 disabled:bg-gray font-bold underline"
-          onClick={async () => {
-            try {
-              setIsRequestingNewCode(true);
-              //no need to inspect result
-              await resendSignUpCode({
-                username: props.codeRecipientEmail.raw,
-              });
-
-              return alert("New code has been sent to your email.");
-            } catch (err) {
-              const message =
-                err instanceof AuthError ? err.message : GENERIC_ERROR_MESSAGE;
-              handleError(message);
-            } finally {
-              setIsRequestingNewCode(false);
-            }
-          }}
-        >
-          00:{String(remainingTime).padStart(2, "0")}
-        </button>
+        <ResendOTPTimer onClick={resendOTP} />
       </span>
     </Form>
   );
