+++ src/components/Signup/ConfirmForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ConfirmForm";
